Data Set and Objective Document for the project
This dataset consists of about 40000 articles consisting of fake as well as real news. Our aim is train our model so that it can correctly predict whether a given piece of news is real or fake. The fake and real news data is given in two separate datasets with each dataset consisting around 20000 articles each.
The whole process and progress of this project is mentioned in the Objective Document and the guide for this project will be Neha madam from @Excelr team.